move ea8b tests to their own module.
